refactor TestHybridSchemeInitialization
